Changed the way to create player and clubs. Both create buttons are NOT WORKING yet. Do it Gusa
